requirements.txt
ftfy
regex
tqdm
seaborn
medmnist
timm
faiss-cpu
scikit-learn
open_clip_torch
transformers